# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BotSync (RAG API)",
        "description": "Документация API для системы BotSync RAG (Retrieval-Augmented Generation), обучение систем на данных для автоматизации чат ботов",
        "contact": {
            "email": "uniqdeveloper@yandex.ru"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Текущий сервер"
        }
    ],
    "paths": {
        "/api/v1/callback": {
            "post": {
                "tags": [
                    "Обучение ассистента на JSON"
                ],
                "summary": "Дообучение ассистента",
                "description": "Этот эндпоинт позволяет отправлять подготовленные данные (вопросы и ответы) напрямую для индексации в базе знаний ассистента.",
                "operationId": "assistantCallback",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assistant_id",
                                    "name",
                                    "chunks"
                                ],
                                "properties": {
                                    "assistant_id": {
                                        "description": "ID ассистента в системе",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "Название источника знаний",
                                        "type": "string",
                                        "example": "FAQ по пиццерии"
                                    },
                                    "chunks": {
                                        "description": "Массив объектов с произвольными данными (например, пары вопрос-ответ)",
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "example": {
                                                "question": "Какую пиццу вы готовите?",
                                                "answer": "Мы готовим большую и тонкую пиццу на дровах."
                                            }
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Данные успешно получены и проиндексированы",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Данные успешно получены и проиндексированы"
                                        },
                                        "assistant_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "knowledge_id": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The assistant id field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "assistant_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The assistant id field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка при индексации данных: [Текст ошибки]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Главная страница панели управления",
                "operationId": "c93acf112f72d6540ff37a56d4de1919",
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            }
        },
        "/share-chat/{id}": {
            "get": {
                "tags": [
                    "Public Chat"
                ],
                "summary": "Публичный чат ассистента",
                "operationId": "32f2e5abfdfd7d88a2f65d98eb482c66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            }
        },
        "/share-chat/{id}/init": {
            "get": {
                "tags": [
                    "Public Chat"
                ],
                "summary": "Инициализация публичного чата",
                "operationId": "41c31e71fa07f22fc962ab2fa23b1b45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные инициализации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "welcome_message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "actions": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "nullable": true
                                        },
                                        "company_name": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "phone": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "description": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "social_networks": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/share-chat/{id}/message": {
            "post": {
                "tags": [
                    "Public Chat"
                ],
                "summary": "Отправка сообщения в публичный чат",
                "operationId": "ba7c22115faf743fb2964f9096915e42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublicMessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ответ ассистента",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "answer": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/connectors": {
            "get": {
                "tags": [
                    "Connectors"
                ],
                "summary": "Список доступных коннекторов",
                "operationId": "918574df35f19f9656bfd3f1ee91bf5e",
                "responses": {
                    "200": {
                        "description": "Успешный ответ"
                    }
                }
            }
        },
        "/connectors/attach-assistant": {
            "post": {
                "tags": [
                    "Connectors"
                ],
                "summary": "Подключение ассистента к коннектору",
                "operationId": "d63be6ac6b0e0c2103908b308ec3c8bb",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "connector_id",
                                    "assistant_id"
                                ],
                                "properties": {
                                    "connector_id": {
                                        "type": "integer"
                                    },
                                    "assistant_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Перенаправление обратно"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AssistantMarketplaceData": {
                "title": "Assistant Marketplace Data",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "category": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "social": {
                        "type": "object",
                        "nullable": true
                    },
                    "welcome_message": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "StoreAssistantRequest": {
                "title": "Store Assistant Request",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "style": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "commercial",
                            "business",
                            "rude",
                            "positive"
                        ]
                    },
                    "brand_name": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 20
                    },
                    "social": {
                        "type": "object",
                        "nullable": true
                    },
                    "fallback": {
                        "type": "string",
                        "nullable": true
                    },
                    "welcome_message": {
                        "type": "string",
                        "nullable": true
                    },
                    "actions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "system": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateAssistantRequest": {
                "title": "Update Assistant Request",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "style": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "commercial",
                            "business",
                            "rude",
                            "positive"
                        ]
                    },
                    "brand_name": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 255
                    },
                    "phone": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 20
                    },
                    "social": {
                        "type": "object",
                        "nullable": true
                    },
                    "fallback": {
                        "type": "string",
                        "nullable": true
                    },
                    "welcome_message": {
                        "type": "string",
                        "nullable": true
                    },
                    "actions": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "system": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PublicMessageRequest": {
                "title": "Public Message Request",
                "required": [
                    "question"
                ],
                "properties": {
                    "question": {
                        "type": "string",
                        "example": "Как дела?",
                        "maxLength": 1000
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Обучение ассистента на JSON",
            "description": "Обучение ассистента на JSON"
        },
        {
            "name": "Dashboard",
            "description": "Dashboard"
        },
        {
            "name": "Public Chat",
            "description": "Public Chat"
        },
        {
            "name": "Connectors",
            "description": "Connectors"
        }
    ]
}